+++ app/src/main/java/com/trelp/aag2020/data/model/MoviesResponse.kt
@@ -5,7 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class MoviesResponse(
-//    @SerialName("dates") val dates: Dates,
     @SerialName("page") val page: Int,
     @SerialName("results") val results: List<Movie>,
     @SerialName("total_pages") val totalPages: Int,
